# contracts/openapi/clm.yaml
...
openapi: 3.1.0
info:
  title: CLM — Claims management API
  version: 1.0.0
  summary: Cross-module and UI/partner operations owned by CLM (PRD-07 §9.1)
  description: 'Contract of the operations CLM exposes (PRD-07 §9.1, canonical names per R-87). Modules call these in-process
    through `CoreIns.<Module>.Contracts` interfaces with the same operation names (D-ARC-16); HTTP is for UI, partners and
    tests. Shared conventions: `common.yaml` and `README.md` (contracts/). Generated from the PRD tables; `x-status: minimal`
    marks operations whose PRD row gives only names, to be completed by the owning work package without inventing rules.'
  x-module: CLM
  x-prd: PRD-07
...
servers:
- url: /
  description: Same origin; paths carry /api/<module>/v<major>
security:
- entra: []
...
paths:
  /api/clm/v1/fnol/submit:
    post:
      operationId: clm.Fnol.submit
      summary: FNOL for all channels
      description: |-
        FNOL for all channels

        PRD inputs: FNOL payload
        PRD outputs: claim, exposures, checklist, assignment
      tags:
      - Fnol
      x-requirement:
      - REQ-CLM-001
      x-anchor:
      - REQ-CLM-001
      x-wave: W7
      x-work-packages:
      - W7-CLM-01
      x-status: full
      x-maturity: pre-release
      x-typed: REQ-CLM-001
      x-operation-kind: command
      x-idempotency: required
      x-dry-run: true
      x-dry-run-note: 'PRD: Yes'
      x-exposure:
      - ui
      - partner
      x-consumers:
      - CHN
      x-partner-routes:
      - CHN facade POST /api/chn/v1/claims
      - MCP tool claim_report_submit
      x-permission: clm.Fnol.submit
      x-error-codes:
      - code: CLM-ERR-FNOL-001
        note: missing field
        status: 422
      - code: CLM-ERR-FNOL-CONSENT
        status: 422
      - code: CLM-ERR-IDEMPOTENCY-MISMATCH
        status: 409
      - code: CLM-ERR-LOSS-DATE
        status: 422
        note: loss date in the future or after the notice date (REQ-CLM-030)
      - code: CLM-ERR-POLICY-UNVERIFIED
        status: 422
        note: POL does not know the policy or gave no snapshot reference (REQ-CLM-002, -050)
      - code: CLM-ERR-DUPLICATE-CANDIDATES
        status: 409
        note: probable duplicates need a duplicateDecision (REQ-CLM-041)
      - code: CLM-ERR-DEPENDENCY-UNAVAILABLE
        status: 503
        note: POL snapshot service unavailable
      x-source:
      - PRD-07 §9.1 l.1466
      parameters:
      - $ref: common.yaml#/components/parameters/IdempotencyKey
      - $ref: common.yaml#/components/parameters/Traceparent
      - $ref: common.yaml#/components/parameters/AcceptLanguage
      - $ref: common.yaml#/components/parameters/DryRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FnolSubmitRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FnolSubmitResponse'
        '400':
          $ref: common.yaml#/components/responses/BadRequest
        '401':
          $ref: common.yaml#/components/responses/Unauthorized
        '403':
          $ref: common.yaml#/components/responses/Forbidden
        '409':
          $ref: common.yaml#/components/responses/Conflict
        '422':
          $ref: common.yaml#/components/responses/UnprocessableContent
        '429':
          $ref: common.yaml#/components/responses/TooManyRequests
        '503':
          $ref: common.yaml#/components/responses/ServiceUnavailable
        '500':
          $ref: common.yaml#/components/responses/InternalError
...
  /api/clm/v1/fnol/{id}:
    get:
      operationId: clm.Fnol.get
      summary: FNOL for all channels
      description: |-
        FNOL for all channels

        PRD inputs: FNOL payload
        PRD outputs: claim, exposures, checklist, assignment
      tags:
      - Fnol
      x-requirement:
      - REQ-CLM-001
      x-anchor:
      - REQ-CLM-001
      x-wave: W7
      x-work-packages:
      - W7-CLM-01
      x-status: full
      x-maturity: pre-release
      x-io-shared: 4
      x-operation-kind: query
      x-idempotency: none
      x-dry-run: false
      x-dry-run-note: 'PRD: Yes'
      x-exposure:
      - ui
      x-permission: clm.Fnol.get
      x-error-codes:
      - code: CLM-ERR-FNOL-001
        note: missing field
        status: 422
      - code: CLM-ERR-FNOL-CONSENT
        status: 422
      - code: CLM-ERR-IDEMPOTENCY-MISMATCH
        status: 409
      - code: CLM-ERR-NOT-FOUND
        status: 404
      x-source:
      - PRD-07 §9.1 l.1466
      parameters:
      - $ref: common.yaml#/components/parameters/Traceparent
      - $ref: common.yaml#/components/parameters/AcceptLanguage
      - name: id
        in: path
        required: true
        description: Identifier of the FNOL snapshot or of its claim (REQ-CLM-044)
        schema:
          type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FnolGetResponse'
        '400':
          $ref: common.yaml#/components/responses/BadRequest
        '401':
          $ref: common.yaml#/components/responses/Unauthorized
        '403':
          $ref: common.yaml#/components/responses/Forbidden
        '404':
          $ref: common.yaml#/components/responses/NotFound
        '409':
          $ref: common.yaml#/components/responses/Conflict
        '422':
          $ref: common.yaml#/components/responses/UnprocessableContent
        '500':
          $ref: common.yaml#/components/responses/InternalError
  /api/clm/v1/fnol/validate:
    post:
      operationId: clm.Fnol.validate
      summary: FNOL for all channels
      description: |-
        FNOL for all channels

        PRD inputs: FNOL payload
        PRD outputs: claim, exposures, checklist, assignment
      tags:
      - Fnol
      x-requirement:
      - REQ-CLM-001
      x-anchor:
      - REQ-CLM-001
      x-wave: W7
      x-work-packages:
      - W7-CLM-01
      x-status: full
      x-maturity: pre-release
      x-io-shared: 4
      x-operation-kind: query
      x-idempotency: none
      x-dry-run: false
      x-dry-run-note: 'PRD: Yes'
      x-exposure:
      - ui
      - partner
      x-consumers:
      - CHN
      x-partner-routes:
      - MCP tool claim_report_draft
      x-permission: clm.Fnol.validate
      x-error-codes:
      - code: CLM-ERR-FNOL-001
        note: missing field
        status: 422
      - code: CLM-ERR-FNOL-CONSENT
        status: 422
      - code: CLM-ERR-IDEMPOTENCY-MISMATCH
        status: 409
      - code: CLM-ERR-DEPENDENCY-UNAVAILABLE
        status: 503
      x-source:
      - PRD-07 §9.1 l.1466
      parameters:
      - $ref: common.yaml#/components/parameters/Traceparent
      - $ref: common.yaml#/components/parameters/AcceptLanguage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FnolValidateRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FnolValidateResponse'
        '400':
          $ref: common.yaml#/components/responses/BadRequest
        '401':
          $ref: common.yaml#/components/responses/Unauthorized
        '403':
          $ref: common.yaml#/components/responses/Forbidden
        '409':
          $ref: common.yaml#/components/responses/Conflict
        '422':
          $ref: common.yaml#/components/responses/UnprocessableContent
        '429':
          $ref: common.yaml#/components/responses/TooManyRequests
        '503':
          $ref: common.yaml#/components/responses/ServiceUnavailable
        '500':
          $ref: common.yaml#/components/responses/InternalError
  /api/clm/v1/claims/{id}:
    get:
      operationId: clm.Claim.get
      summary: Claim lifecycle
      description: |-
        Claim lifecycle

        PRD inputs: ids, changes, reason
        PRD outputs: claim
      tags:
      - Claim
      x-requirement:
      - REQ-CLM-061
      - REQ-CLM-071..REQ-CLM-074
      - REQ-CLM-042
      x-wave: W7
      x-work-packages:
      - W7-CLM-01
      x-status: full
      x-maturity: pre-release
      x-io-shared: 5
      x-operation-kind: query
      x-idempotency: none
      x-dry-run: false
      x-dry-run-note: 'PRD dry-run: Yes (close, reopen); PRD idempotency: Yes'
      x-exposure:
      - ui
      x-consumers:
      - RI
      x-permission: clm.Claim.get
      x-error-codes:
      - code: CLM-ERR-ILLEGAL-TRANSITION
        status: 422
      - code: CLM-ERR-CLOSE-GUARD
        status: 422
      - code: CLM-ERR-NOT-FOUND
        status: 404
      x-source:
      - PRD-07 §9.1 l.1467
      parameters:
      - $ref: common.yaml#/components/parameters/Traceparent
      - $ref: common.yaml#/components/parameters/AcceptLanguage
      - name: id
        in: path
        required: true
        description: 'Identifier of the Claim (PRD input: "ids")'
        schema:
          type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimGetResponse'
        '400':
          $ref: common.yaml#/components/responses/BadRequest
        '401':
          $ref: common.yaml#/components/responses/Unauthorized
        '403':
          $ref: common.yaml#/components/responses/Forbidden
        '404':
          $ref: common.yaml#/components/responses/NotFound
        '422':
          $ref: common.yaml#/components/responses/UnprocessableContent
        '500':
          $ref: common.yaml#/components/responses/InternalError
    patch:
      operationId: clm.Claim.update
      summary: Claim lifecycle
      description: |-
        Claim lifecycle

        PRD inputs: ids, changes, reason
        PRD outputs: claim
      tags:
      - Claim
      x-requirement:
      - REQ-CLM-061
      - REQ-CLM-071..REQ-CLM-074
      - REQ-CLM-042
      x-wave: W7
      x-work-packages:
      - W7-CLM-01
      x-status: full
      x-maturity: pre-release
      x-io-shared: 5
      x-operation-kind: command
      x-idempotency: required
      x-dry-run: false
      x-dry-run-note: 'PRD dry-run: Yes (close, reopen); PRD idempotency: Yes'
      x-exposure:
      - ui
      x-permission: clm.Claim.update
      x-authority-types:
      - CLM.REOPEN
      x-authority-source:
      - PRD-07 l.504
      x-error-codes:
      - code: CLM-ERR-ILLEGAL-TRANSITION
        status: 422
      - code: CLM-ERR-CLOSE-GUARD
        status: 422
      - code: CLM-ERR-IDEMPOTENCY-MISMATCH
        note: contract §3.5.3 (replay with a different payload)
        status: 409
      x-source:
      - PRD-07 §9.1 l.1467
      parameters:
      - $ref: common.yaml#/components/parameters/IdempotencyKey
      - $ref: common.yaml#/components/parameters/Traceparent
      - $ref: common.yaml#/components/parameters/AcceptLanguage
      - name: id
        in: path
        required: true
        description: Identifier of the Claim
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimUpdateRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimUpdateResponse'
        '400':
          $ref: common.yaml#/components/responses/BadRequest
        '401':
          $ref: common.yaml#/components/responses/Unauthorized
        '403':
          $ref: common.yaml#/components/responses/Forbidden
        '404':
          $ref: common.yaml#/components/responses/NotFound
        '409':
          $ref: common.yaml#/components/responses/Conflict
        '422':
          $ref: common.yaml#/components/responses/UnprocessableContent
        '500':
          $ref: common.yaml#/components/responses/InternalError
  /api/clm/v1/claims/close:
    post:
      operationId: clm.Claim.close
      summary: Claim lifecycle
      description: |-
        Claim lifecycle

        PRD inputs: ids, changes, reason
        PRD outputs: claim
      tags:
      - Claim
      x-requirement:
      - REQ-CLM-061
      - REQ-CLM-071..REQ-CLM-074
      - REQ-CLM-042
      x-wave: W7
      x-work-packages:
      - W7-CLM-01
      x-status: full
      x-maturity: pre-release
      x-io-shared: 5
      x-operation-kind: command
      x-idempotency: required
      x-dry-run: true
      x-dry-run-note: 'PRD dry-run: Yes (close, reopen); PRD idempotency: Yes'
      x-exposure:
      - ui
      x-permission: clm.Claim.close
      x-authority-types:
      - CLM.REOPEN
      x-authority-source:
      - PRD-07 l.504
      x-error-codes:
      - code: CLM-ERR-ILLEGAL-TRANSITION
        status: 422
      - code: CLM-ERR-CLOSE-GUARD
        status: 422
      - code: CLM-ERR-IDEMPOTENCY-MISMATCH
        note: contract §3.5.3 (replay with a different payload)
        status: 409
      - code: CLM-ERR-STALE
        status: 409
        note: the claim changed meanwhile (record version)
      - code: CLM-ERR-NOT-FOUND
        status: 404
      x-source:
      - PRD-07 §9.1 l.1467
      parameters:
      - $ref: common.yaml#/components/parameters/IdempotencyKey
      - $ref: common.yaml#/components/parameters/Traceparent
      - $ref: common.yaml#/components/parameters/AcceptLanguage
      - $ref: common.yaml#/components/parameters/DryRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimCloseRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimCloseResponse'
        '400':
          $ref: common.yaml#/components/responses/BadRequest
        '401':
          $ref: common.yaml#/components/responses/Unauthorized
        '403':
          $ref: common.yaml#/components/responses/Forbidden
        '409':
          $ref: common.yaml#/components/responses/Conflict
        '422':
          $ref: common.yaml#/components/responses/UnprocessableContent
        '404':
          $ref: common.yaml#/components/responses/NotFound
        '500':
          $ref: common.yaml#/components/responses/InternalError
...
  /api/clm/v1/claims/search:
    get:
      operationId: clm.Claim.search
      summary: Claim search (R-07)
      description: |-
        Claim search (R-07)

        PRD inputs: criteria
        PRD outputs: page of claims
      tags:
      - Claim
      x-requirement:
      - REQ-CLM-011
      x-anchor:
      - REQ-CLM-011
      x-wave: W7
      x-work-packages:
      - W7-CLM-05
      x-status: full
      x-maturity: pre-release
      x-operation-kind: query
      x-idempotency: none
      x-dry-run: false
      x-dry-run-note: 'PRD: n/a'
      x-exposure:
      - ui
      x-consumers:
      - CMP
      - WRK
      x-permission: clm.Claim.search
      x-error-codes:
      - code: CLM-ERR-SEARCH-CRITERIA
        status: 422
      x-source:
      - PRD-07 §9.1 l.1468
      parameters:
      - $ref: common.yaml#/components/parameters/Traceparent
      - $ref: common.yaml#/components/parameters/AcceptLanguage
      - $ref: common.yaml#/components/parameters/Cursor
      - $ref: common.yaml#/components/parameters/Limit
      - name: claimNumber
        in: query
        required: false
        description: Claim number (P0). Every other criterion, personal or not, goes in the body of the POST form (D-SLC-05)
        schema:
          type: string
          maxLength: 64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimSearchPage'
        '400':
          $ref: common.yaml#/components/responses/BadRequest
        '401':
          $ref: common.yaml#/components/responses/Unauthorized
        '403':
          $ref: common.yaml#/components/responses/Forbidden
        '422':
          $ref: common.yaml#/components/responses/UnprocessableContent
        '500':
          $ref: common.yaml#/components/responses/InternalError
    post:
      operationId: clm.Claim.searchByCriteria
      summary: Claim search (POST form, criteria in the body)
      description: |-
        POST form of clm.Claim.search (D-SLC-05): policy number and party ids travel in the body, never in a URL. A read;
        no Idempotency-Key. Results never show free text and never count claims the caller may not see (REQ-CLM-011).

        PRD inputs: criteria
        PRD outputs: page of claims
      tags:
      - Claim
      x-requirement:
      - REQ-CLM-011
      x-wave: W7
      x-work-packages:
      - SL2-CLM-CORE
      x-status: full
      x-maturity: pre-release
      x-operation-kind: query
      x-idempotency: none
      x-dry-run: false
      x-dry-run-note: 'PRD: n/a'
      x-exposure:
      - ui
      x-consumers:
      - CMP
      - WRK
      x-permission: clm.Claim.search
      x-prd-name: clm.Claim.search
      x-error-codes:
      - code: CLM-ERR-SEARCH-CRITERIA
        status: 422
      x-source:
      - PRD-07 §9.1 l.1468
      - DECISIONS D-SLC-05
      parameters:
      - $ref: common.yaml#/components/parameters/Traceparent
      - $ref: common.yaml#/components/parameters/AcceptLanguage
      - $ref: common.yaml#/components/parameters/Cursor
      - $ref: common.yaml#/components/parameters/Limit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimSearchCriteria'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimSearchPage'
        '400':
          $ref: common.yaml#/components/responses/BadRequest
        '401':
          $ref: common.yaml#/components/responses/Unauthorized
        '403':
          $ref: common.yaml#/components/responses/Forbidden
        '422':
          $ref: common.yaml#/components/responses/UnprocessableContent
        '500':
          $ref: common.yaml#/components/responses/InternalError
  /api/clm/v1/exposures:
    post:
      operationId: clm.Exposure.create
      summary: Exposures
      description: |-
        Exposures

        PRD inputs: exposure data
        PRD outputs: exposure
      tags:
      - Exposure
      x-requirement:
      - REQ-CLM-062
      - REQ-CLM-063
      - REQ-CLM-072
      x-wave: W7
      x-work-packages:
      - W7-CLM-01
      x-status: full
      x-maturity: pre-release
      x-io-shared: 4
      x-operation-kind: command
      x-idempotency: required
      x-dry-run: true
      x-dry-run-note: 'PRD: Yes'
      x-exposure:
      - ui
      x-permission: clm.Exposure.create
      x-error-codes:
      - code: CLM-ERR-EXPOSURE-DUPLICATE
        status: 409
      - code: CLM-ERR-IDEMPOTENCY-MISMATCH
        note: contract §3.5.3 (replay with a different payload)
        status: 409
      - code: CLM-ERR-ILLEGAL-TRANSITION
        status: 422
        note: claim not Open (REQ-CLM-071)
      - code: CLM-ERR-STALE
        status: 409
        note: the claim changed meanwhile (record version)
      - code: CLM-ERR-NOT-FOUND
        status: 404
      x-source:
      - PRD-07 §9.1 l.1469
      parameters:
      - $ref: common.yaml#/components/parameters/IdempotencyKey
      - $ref: common.yaml#/components/parameters/Traceparent
      - $ref: common.yaml#/components/parameters/AcceptLanguage
      - $ref: common.yaml#/components/parameters/DryRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExposureCreateRequest'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExposureCreateResponse'
        '400':
          $ref: common.yaml#/components/responses/BadRequest
        '401':
          $ref: common.yaml#/components/responses/Unauthorized
        '403':
          $ref: common.yaml#/components/responses/Forbidden
        '409':
          $ref: common.yaml#/components/responses/Conflict
        '422':
          $ref: common.yaml#/components/responses/UnprocessableContent
        '404':
          $ref: common.yaml#/components/responses/NotFound
        '500':
          $ref: common.yaml#/components/responses/InternalError
...
  /api/clm/v1/redresses/request:
    post:
      operationId: clm.Redress.request
      summary: Claim-related complaint redress from CMP paid as ex-gratia
      description: |-
        Claim-related complaint redress from CMP paid as ex-gratia

        PRD inputs: complaint id, claim id, payee, amount, reason
        PRD outputs: set id, status
      tags:
      - Redress
      x-requirement:
      - REQ-CLM-265
      x-wave: unscheduled
      x-status: full
      x-maturity: pre-release
      x-operation-kind: command
      x-idempotency: required
      x-dry-run: true
      x-dry-run-note: 'PRD: Yes'
      x-exposure:
      - ui
      x-consumers:
      - CMP
      x-permission: clm.Redress.request
      x-authority-types:
      - CLM.EXGRATIA
      x-authority-source:
      - PRD-07 l.593
      x-error-codes:
      - code: CLM-ERR-NOT-PAYABLE
        status: 422
      - code: CLM-ERR-AUTHORITY
        status: 403
      - code: CLM-ERR-IDEMPOTENCY-MISMATCH
        note: contract §3.5.3 (replay with a different payload)
        status: 409
      x-source:
      - PRD-07 §9.1 l.1473
      parameters:
      - $ref: common.yaml#/components/parameters/IdempotencyKey
      - $ref: common.yaml#/components/parameters/Traceparent
      - $ref: common.yaml#/components/parameters/AcceptLanguage
      - $ref: common.yaml#/components/parameters/DryRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedressRequestRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedressRequestResponse'
        '400':
          $ref: common.yaml#/components/responses/BadRequest
        '401':
          $ref: common.yaml#/components/responses/Unauthorized
        '403':
          $ref: common.yaml#/components/responses/Forbidden
        '409':
          $ref: common.yaml#/components/responses/Conflict
        '422':
          $ref: common.yaml#/components/responses/UnprocessableContent
        '500':
          $ref: common.yaml#/components/responses/InternalError
...
  /api/clm/v1/certificates/request:
    post:
      operationId: clm.Certificate.request
      summary: Claims-history certificate
      description: |-
        Claims-history certificate

        PRD inputs: requester, policies, period
        PRD outputs: certificate
      tags:
      - Certificate
      x-requirement:
      - REQ-CLM-008
      - REQ-CLM-170..REQ-CLM-173
      x-anchor:
      - REQ-CLM-008
      x-wave: W7
      x-work-packages:
      - W7-CLM-04
      x-status: full
      x-maturity: pre-release
      x-io-shared: 3
      x-operation-kind: command
      x-idempotency: required
      x-dry-run: true
      x-dry-run-note: 'PRD: Yes'
      x-exposure:
      - ui
      x-consumers:
      - CHN
      x-permission: clm.Certificate.request
      x-error-codes:
      - code: CLM-ERR-CERT-REQUESTER
        status: 422
      - code: CLM-ERR-IDEMPOTENCY-MISMATCH
        note: contract §3.5.3 (replay with a different payload)
        status: 409
      x-source:
      - PRD-07 §9.1 l.1480
      parameters:
      - $ref: common.yaml#/components/parameters/IdempotencyKey
      - $ref: common.yaml#/components/parameters/Traceparent
      - $ref: common.yaml#/components/parameters/AcceptLanguage
      - $ref: common.yaml#/components/parameters/DryRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertificateRequestRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificateRequestResponse'
        '400':
          $ref: common.yaml#/components/responses/BadRequest
        '401':
          $ref: common.yaml#/components/responses/Unauthorized
        '403':
          $ref: common.yaml#/components/responses/Forbidden
        '409':
          $ref: common.yaml#/components/responses/Conflict
        '422':
          $ref: common.yaml#/components/responses/UnprocessableContent
        '500':
          $ref: common.yaml#/components/responses/InternalError
...
components:
  securitySchemes:
    entra:
      $ref: common.yaml#/components/securitySchemes/entra
  schemas:
...
    FnolSubmitResponse:
      type: object
      properties:
        claimId:
          $ref: ../events/common.schema.json#/$defs/Uuid
        claimNumber:
          $ref: ../events/common.schema.json#/$defs/BusinessNumber
        claim:
          $ref: '#/components/schemas/ClaimSummary'
        exposures:
          type: array
          items:
            $ref: '#/components/schemas/ExposureView'
        coverageIndications:
          type: array
          items:
            $ref: '#/components/schemas/CoverageIndication'
        duplicateCandidates:
          type: array
          items:
            $ref: '#/components/schemas/DuplicateCandidate'
        handlingSegment:
          $ref: ../events/common.schema.json#/$defs/Code
        assignment:
          type: object
          properties:
            assigneeType:
              type: string
              enum:
              - USER
              - GROUP
              - QUEUE
              description: WRK assignee kind
            assigneeId:
              type: string
        requiredDocuments:
          type: array
          items:
            $ref: ../events/common.schema.json#/$defs/Code
      required:
      - claimId
      - claimNumber
      - claim
      - exposures
      - coverageIndications
      - duplicateCandidates
      - handlingSegment
      description: 'Typed from REQ-CLM-001 (SL2-CLM-CORE). PRD outputs: "claim, exposures, checklist, assignment". Holds no free text (stored by the idempotency store).'
...
    FnolValidateRequest:
      type: object
      description: 'clm.Fnol.validate request: the FNOL payload a submit would send (no claim, number or event is created).'
      properties:
        fnol:
          $ref: '#/components/schemas/FnolSubmitRequest'
      required:
      - fnol
...
    ClaimGetResponse:
      type: object
      description: 'clm.Claim.get result. PRD outputs: "claim"'
      properties:
        claim:
          $ref: '#/components/schemas/ClaimView'
      required:
      - claim
...